const_crypto

Module ed25519

Magnetar Fields:

Most of the code in here and the modules within were copied from curve curve25519-dalek and adapted to be const. Since the intended use case of is_on_curve is not intended to be used with any secrets, there were constant time operations that were replaced with unsafer counterparts. As such, this is not a cryptographically safe implementation. It is only intended to be used at compile time with public keys. For the sake of being as uninvasive as possible, there are some relic constant time implementations for some operations, and there may be some misnamed functions.

There is a test in this module which checks that the is_on_curve evaluation agrees with a large batch of random keys.

Structs

CompressedEdwardsY

In “Edwards y” / “Ed25519” format, the curve point \((x,y)\) is determined by the \(y\)-coordinate and the sign of \(x\).

EdwardsPoint

PartialPda

Functions

crypto_unsafe_is_on_curve

Do not use as part of a protocol that deals with secrets! Only use to evaluate if

derive_program_address